Added policies for advertisements

+++ app/Http/Controllers/Advertisement/DestroyController.php
@@ -8,6 +8,8 @@ class DestroyController extends BaseController
 {
     public function __invoke(Advertisement $advertisement)
     {
+        $this->authorize('delete', $advertisement);
+
         $this->service->destroy($advertisement);
 
         return redirect()->route('advertisement.index');
